Wextra fixes

+++ src/buffer.c
@@ -143,8 +143,8 @@ buffer_set_navail_locked(const struct buffer_set *bs)
 
     NIOVA_ASSERT(bs->bs_num_allocated  >= 0);
     NIOVA_ASSERT(bs->bs_num_pndg_alloc >= 0);
-    NIOVA_ASSERT(bs->bs_num_bufs >= (bs->bs_num_allocated +
-                                     bs->bs_num_pndg_alloc));
+    NIOVA_ASSERT((ssize_t)bs->bs_num_bufs >=
+                 (bs->bs_num_allocated + bs->bs_num_pndg_alloc));
 
     return bs->bs_num_bufs - (bs->bs_num_allocated + bs->bs_num_pndg_alloc);
 }
@@ -215,7 +215,7 @@ buffer_set_allocate_item_locked(struct buffer_set *bs)
     bs->bs_num_allocated++;
     bi->bi_allocated = true;
 
-    if (bs->bs_num_allocated > bs->bs_max_allocated)
+    if (bs->bs_num_allocated > (ssize_t)bs->bs_max_allocated)
         bs->bs_max_allocated = bs->bs_num_allocated;
 
     return bi;
@@ -262,7 +262,7 @@ buffer_set_release_pending_alloc(struct buffer_set *bs, const size_t nitems)
 
     BS_LOCK(bs);
 
-    if (nitems > bs->bs_num_pndg_alloc)
+    if ((ssize_t)nitems > bs->bs_num_pndg_alloc)
     {
         BS_UNLOCK(bs);
